+++ CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 3.5.0 - 2021-02-05
+
+- Changed `displayName` of `slack_user` to use `display_name` or `real_name` or
+  `name`, fallback to `id` only when those are undefined.
+
+- Changed `username` property to use value from `user.name` instead of
+  `user.id`.
+
+- Added `userId` property using value from `user.id`.
+
+- Added `admin` boolean property to `slack_user`, as it is a normalized property
+  on the `User` class entity.
+
+- Added normalized boolean properties `active`, `archived`, `public`, `private`
+  to the `slack_channel` entity.
+
 ## 3.4.1 - 2020-11-24
 
 - Added retries for `slack_webapi_platform_error` error codes. The Slack Client